requirements.txt
click==6.6
Flask==0.11.1
Flask-Classy==0.6.10
Flask-HTTPAuth==3.1.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
py-enigma==0.1
py-enigma-operator==0.6
pygal==2.2.3
pytz==2016.6
tzlocal==1.2.2
Werkzeug==0.11.10
